WIP: partial FFMPEG based calculations

+++ pkg/cue/result.go
@@ -8,11 +8,11 @@ import (
 
 // Result - CUE calculations data
 type Result struct {
-	Duration              float32 `json:"duration" yaml:"duration"`
-	CueDuration           float32 `json:"liq_cue_duration" yaml:"liq_cue_duration"`
-	CueIn                 float32 `json:"liq_cue_in" yaml:"liq_cue_in"`
-	CueOut                float32 `json:"liq_cue_out" yaml:"liq_cue_out"`
-	CrossStartNext        float32 `json:"liq_cross_start_next" yaml:"liq_cross_start_next"`
+	Duration              float64 `json:"duration" yaml:"duration"`
+	CueDuration           float64 `json:"liq_cue_duration" yaml:"liq_cue_duration"`
+	CueIn                 float64 `json:"liq_cue_in" yaml:"liq_cue_in"`
+	CueOut                float64 `json:"liq_cue_out" yaml:"liq_cue_out"`
+	CrossStartNext        float64 `json:"liq_cross_start_next" yaml:"liq_cross_start_next"`
 	LongTail              bool    `json:"liq_longtail" yaml:"liq_longtail"`
 	SustainedLoudnessDrop bool    `json:"liq_sustained_ending" yaml:"liq_sustained_ending"`
 	Loudness              string  `json:"liq_loudness" yaml:"liq_loudness"`
@@ -20,9 +20,9 @@ type Result struct {
 	Amplify               string  `json:"liq_amplify" yaml:"liq_amplify"`
 	AmplifyAdjustment     string  `json:"liq_amplify_adjustment" yaml:"liq_amplify_adjustment"`
 	ReferenceLoudness     string  `json:"liq_reference_loudness" yaml:"liq_reference_loudness"`
-	BlankSkip             float32 `json:"liq_blankskip" yaml:"liq_blankskip"`
+	BlankSkip             float64 `json:"liq_blankskip" yaml:"liq_blankskip"`
 	BlankSkipped          bool    `json:"liq_blank_skipped" yaml:"liq_blank_skipped"`
-	TruePeak              float32 `json:"liq_true_peak" yaml:"liq_true_peak"`
+	TruePeak              float64 `json:"liq_true_peak" yaml:"liq_true_peak"`
 	TruePeakDb            string  `json:"liq_true_peak_db" yaml:"liq_true_peak_db"`
 }
